poco178: remove costmap_converter
Poco1.7.8 causes a segfault which seems to come when a library that is
to be dynamically loaded was already linked against the loading library.
This scenario is the case between teb_local_planner and
costmap_converter.

This and the following commits are supposed to be used as a workaround
in order to get a working teb_local_planner again. It is intended to
fix the issue though.